+++ lib/pathplan/pathplan.h
@@ -24,6 +24,7 @@
 #pragma once
 
 #include "pathgeom.h"
+#include <stddef.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -46,10 +47,10 @@ extern "C" {
 			     Ppolyline_t * output_route);
 
 /* fit a spline to an input polyline, without touching barrier segments */
-    PATHPLAN_API int Proutespline(Pedge_t * barriers, int n_barriers,
-			    Ppolyline_t input_route,
-			    Pvector_t endpoint_slopes[2],
-			    Ppolyline_t * output_route);
+PATHPLAN_API int Proutespline(Pedge_t *barriers, size_t n_barriers,
+                              Ppolyline_t input_route,
+                              Pvector_t endpoint_slopes[2],
+                              Ppolyline_t *output_route);
 
 /* utility function to convert from a set of polygonal obstacles to barriers */
     PATHPLAN_API int Ppolybarriers(Ppoly_t ** polys, int npolys,
